+++ Directory.Build.props
@@ -3,7 +3,7 @@
 <Project>
 
 	<PropertyGroup>
-		<Version>1.6.5</Version>
+		<Version>1.6.6</Version>
 		<TargetFrameworks>net6.0;net7.0;net8.0</TargetFrameworks>
 		<LangVersion>preview</LangVersion>
 		<RepositoryUrl>https://github.com/EasilyNET/EasilyNET</RepositoryUrl>
