+++ EasilyNET.sln
@@ -54,9 +54,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "EasilyNET.Mongo.Core", "src
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "EasilyNET.MongoGridFS.AspNetCore", "src\EasilyNET.MongoGridFS.AspNetCore\EasilyNET.MongoGridFS.AspNetCore.csproj", "{5DE016FE-94B3-4F71-9170-35728F09A710}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "EasilyNET.PropertyInjection", "src\EasilyNET.PropertyInjection\EasilyNET.PropertyInjection.csproj", "{02AB99C2-FDF3-48AF-9E51-E6C9A7E67DFD}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "EasilyNET.PropertyInjection.AspNetCore", "src\EasilyNET.PropertyInjection.AspNetCore\EasilyNET.PropertyInjection.AspNetCore.csproj", "{D4D4367E-54A7-4A0B-B41A-CC3D4DD40756}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EasilyNET.AutoDependencyInjection.Core", "src\EasilyNET.AutoDependencyInjection.Core\EasilyNET.AutoDependencyInjection.Core.csproj", "{AC36AE56-981E-4910-9D1C-DD17D4EC28AD}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -128,14 +126,10 @@ Global
 		{5DE016FE-94B3-4F71-9170-35728F09A710}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{5DE016FE-94B3-4F71-9170-35728F09A710}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{5DE016FE-94B3-4F71-9170-35728F09A710}.Release|Any CPU.Build.0 = Release|Any CPU
-		{02AB99C2-FDF3-48AF-9E51-E6C9A7E67DFD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{02AB99C2-FDF3-48AF-9E51-E6C9A7E67DFD}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{02AB99C2-FDF3-48AF-9E51-E6C9A7E67DFD}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{02AB99C2-FDF3-48AF-9E51-E6C9A7E67DFD}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D4D4367E-54A7-4A0B-B41A-CC3D4DD40756}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D4D4367E-54A7-4A0B-B41A-CC3D4DD40756}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D4D4367E-54A7-4A0B-B41A-CC3D4DD40756}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D4D4367E-54A7-4A0B-B41A-CC3D4DD40756}.Release|Any CPU.Build.0 = Release|Any CPU
+		{AC36AE56-981E-4910-9D1C-DD17D4EC28AD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{AC36AE56-981E-4910-9D1C-DD17D4EC28AD}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{AC36AE56-981E-4910-9D1C-DD17D4EC28AD}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{AC36AE56-981E-4910-9D1C-DD17D4EC28AD}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -160,8 +154,7 @@ Global
 		{1FDD0D1C-991B-45C4-8BB8-E9E27CA53082} = {2CB52E13-C7D0-4C8C-BFAD-F1A115AD5373}
 		{F096FF03-E517-4F1D-8657-7946A5DDC06F} = {6B734C3C-FE4C-405D-80E3-AFEAD63BCF2D}
 		{5DE016FE-94B3-4F71-9170-35728F09A710} = {6B734C3C-FE4C-405D-80E3-AFEAD63BCF2D}
-		{02AB99C2-FDF3-48AF-9E51-E6C9A7E67DFD} = {BD6D1E2B-691E-4E8B-9E8D-E7FD2EDDF1F8}
-		{D4D4367E-54A7-4A0B-B41A-CC3D4DD40756} = {BD6D1E2B-691E-4E8B-9E8D-E7FD2EDDF1F8}
+		{AC36AE56-981E-4910-9D1C-DD17D4EC28AD} = {BD6D1E2B-691E-4E8B-9E8D-E7FD2EDDF1F8}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {04F52789-9FF4-4AB9-9132-1789A00E7E4A}
