+++ README.md
@@ -36,7 +36,6 @@ EasilyNET Packages
 | --------------------------------------------------------------------------------------------------------------- | ------------------------------------------------------------------------------- | -------------------------------------------------------------------------------- | --------------------------------------------- |
 | [EasilyNET.AutoDependencyInjection](https://www.nuget.org/packages/EasilyNET.AutoDependencyInjection)           | ![Nuget](https://img.shields.io/nuget/v/EasilyNET.AutoDependencyInjection)      | ![Nuget](https://img.shields.io/nuget/dt/EasilyNET.AutoDependencyInjection)      | 模块化自动注入服务                            |
 | [EasilyNET.Images](https://www.nuget.org/packages/EasilyNET.Images)                                             | ![Nuget](https://img.shields.io/nuget/v/EasilyNET.Images)                       | ![Nuget](https://img.shields.io/nuget/dt/EasilyNET.Images)                       | 一些涉及到图形的工具包,目前仅有 QrCode        |
-| [EasilyNET.PropertyInjection.AspNetCore](https://www.nuget.org/packages/EasilyNET.PropertyInjection.AspNetCore) | ![Nuget](https://img.shields.io/nuget/v/EasilyNET.PropertyInjection.AspNetCore) | ![Nuget](https://img.shields.io/nuget/dt/EasilyNET.PropertyInjection.AspNetCore) | 属性注入                                      |
 | [EasilyNET.RabbitBus.AspNetCore](https://www.nuget.org/packages/EasilyNET.RabbitBus.AspNetCore)                 | ![Nuget](https://img.shields.io/nuget/v/EasilyNET.RabbitBus.AspNetCore)         | ![Nuget](https://img.shields.io/nuget/dt/EasilyNET.RabbitBus.AspNetCore)         | 基于 RabbitMQ 的消息总线处理方案              |
 | [EasilyNET.Security](https://www.nuget.org/packages/EasilyNET.Security)                                         | ![Nuget](https://img.shields.io/nuget/v/EasilyNET.Security)                     | ![Nuget](https://img.shields.io/nuget/dt/EasilyNET.Security)                     | 一个常用加密算法的封装库,从使用简单的目的出发 |
 
@@ -47,9 +46,8 @@ EasilyNET Packages
 | [EasilyNET.IdentityServer.MongoStorage](https://www.nuget.org/packages/EasilyNET.IdentityServer.MongoStorage) | ![Nuget](https://img.shields.io/nuget/v/EasilyNET.IdentityServer.MongoStorage) | ![Nuget](https://img.shields.io/nuget/dt/EasilyNET.IdentityServer.MongoStorage) | IDS6.x 的 Mongodb 持久化支持方案                      |
 | [EasilyNET.Mongo.AspNetCore](https://www.nuget.org/packages/EasilyNET.Mongo.AspNetCore)                       | ![Nuget](https://img.shields.io/nuget/v/EasilyNET.Mongo.AspNetCore)            | ![Nuget](https://img.shields.io/nuget/dt/EasilyNET.Mongo.AspNetCore)            | MongoDB 驱动扩展                                      |
 | [EasilyNET.Mongo.ConsoleDebug](https://www.nuget.org/packages/EasilyNET.Mongo.ConsoleDebug)                   | ![Nuget](https://img.shields.io/nuget/v/EasilyNET.Mongo.ConsoleDebug)          | ![Nuget](https://img.shields.io/nuget/dt/EasilyNET.Mongo.ConsoleDebug)          | MongoDB 的执行命令输出到控制台                        |
-| [EasilyNET.Mongo.Core](https://www.nuget.org/packages/EasilyNET.Mongo.Core)                                   | ![Nuget](https://img.shields.io/nuget/v/EasilyNET.Mongo.Core)                  | ![Nuget](https://img.shields.io/nuget/dt/EasilyNET.Mongo.Core)                  | MongoDB 驱动基础类型支持,便于将业务代码和配置代码分开 |
-| [EasilyNET.MongoSerializer.AspNetCore](https://www.nuget.org/packages/EasilyNET.MongoSerializer.AspNetCore)   | ![Nuget](https://img.shields.io/nuget/v/EasilyNET.MongoSerializer.AspNetCore)  | ![Nuget](https://img.shields.io/nuget/dt/EasilyNET.MongoSerializer.AspNetCore)  | MongoDB 的类型扩展,以及自定义类型扩展方案             |
 | [EasilyNET.MongoGridFS.AspNetCore](https://www.nuget.org/packages/EasilyNET.MongoGridFS.AspNetCore)           | ![Nuget](https://img.shields.io/nuget/v/EasilyNET.MongoGridFS.AspNetCore)      | ![Nuget](https://img.shields.io/nuget/dt/EasilyNET.MongoGridFS.AspNetCore)      | MongoDB GridFS 对象存储解决方案,使对象存储操作简便    |
+| [EasilyNET.MongoSerializer.AspNetCore](https://www.nuget.org/packages/EasilyNET.MongoSerializer.AspNetCore)   | ![Nuget](https://img.shields.io/nuget/v/EasilyNET.MongoSerializer.AspNetCore)  | ![Nuget](https://img.shields.io/nuget/dt/EasilyNET.MongoSerializer.AspNetCore)  | MongoDB 的类型扩展,以及自定义类型扩展方案             |
 
 ![微信公众号](https://github.com/joesdu/joesdu/blob/main/wechat-official-account.png#pic_center)
