+++ Test/WebApi.Test.Unit/AppWebModule.cs
@@ -1,6 +1,5 @@
 ﻿using EasilyNET.AutoDependencyInjection.Attributes;
 using EasilyNET.AutoDependencyInjection.Contexts;
-using EasilyNET.AutoDependencyInjection.Extensions;
 using EasilyNET.AutoDependencyInjection.Modules;
 using EasilyNET.WebCore.Middleware;
