+++ Test/WebApi.Test.Unit/EventHandlers/RabbitEventHandlers.cs
@@ -1,4 +1,4 @@
-﻿using EasilyNET.AutoDependencyInjection.Attributes;
+﻿using EasilyNET.AutoDependencyInjection.Core.Attributes;
 using EasilyNET.RabbitBus.Core;
 using System.Text.Json;
 using WebApi.Test.Unit.Events;
