+++ Test/WebApi.Test.Unit/ServiceModules/SwaggerModule.cs
@@ -1,7 +1,5 @@
 ﻿using EasilyNET.AutoDependencyInjection.Contexts;
-using EasilyNET.AutoDependencyInjection.Extensions;
 using EasilyNET.AutoDependencyInjection.Modules;
-using EasilyNET.WebCore.Swagger;
 using Microsoft.OpenApi.Models;
 
 namespace WebApi.Test.Unit;
