+++ Test/WebApi.Test.Unit/WebApi.Test.Unit.csproj
@@ -22,7 +22,6 @@
         <ProjectReference Include="..\..\src\EasilyNET.MongoSerializer.AspNetCore\EasilyNET.MongoSerializer.AspNetCore.csproj" />
         <ProjectReference Include="..\..\src\EasilyNET.MongoGridFS.AspNetCore\EasilyNET.MongoGridFS.AspNetCore.csproj" />
         <ProjectReference Include="..\..\src\EasilyNET.Mongo.AspNetCore\EasilyNET.Mongo.AspNetCore.csproj" />
-        <ProjectReference Include="..\..\src\EasilyNET.PropertyInjection.AspNetCore\EasilyNET.PropertyInjection.AspNetCore.csproj" />
         <ProjectReference Include="..\..\src\EasilyNET.RabbitBus.AspNetCore\EasilyNET.RabbitBus.AspNetCore.csproj" />
         <ProjectReference Include="..\..\src\EasilyNET.WebCore.Swagger\EasilyNET.WebCore.Swagger.csproj" />
     </ItemGroup>
