+++ Test/WebApi.Test.Unit/appsettings.json
@@ -10,9 +10,5 @@
     "Mongo": "mongodb://guest:guest@127.0.1:27018/?authSource=admin&serverSelectionTimeoutMS=1000",
     // 参考链接: https://www.rabbitmq.com/uri-spec.html
     "Rabbit": "amqp://guest:guest@127.0.0.1:5672/%2f"
-  },
-  "EasilyFS": {
-    "VirtualPath": "/fs",
-    "PhysicalPath": "D:\\test"
   }
 }
