+++ src/EasilyNET.AutoDependencyInjection.Core/EasilyNET.AutoDependencyInjection.Core.csproj
@@ -1,0 +1,12 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+	  <PackageProjectUrl>https://www.nuget.org/packages/EasilyNET.AutoDependencyInjection.Core</PackageProjectUrl>
+	  <Description>自动注入和属性注入以及模块化核心库</Description>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="8.0.0-preview.6.23329.7" />
+  </ItemGroup>
+
+</Project>
