+++ src/EasilyNET.AutoDependencyInjection.Core/README.md
@@ -1,0 +1,3 @@
+﻿#### EasilyNET.AutoDependencyInjection.Core
+
+EasilyNET.AutoDependencyInjection核心库,需要结合EasilyNET.AutoDependencyInjection使用.
