--- src/EasilyNET.AutoDependencyInjection/Extensions/ApplicationInitializationExtensions.cs
@@ -1,20 +1,0 @@
-﻿using EasilyNET.AutoDependencyInjection.Abstractions;
-using EasilyNET.AutoDependencyInjection.Contexts;
-using Microsoft.AspNetCore.Builder;
-using Microsoft.Extensions.DependencyInjection;
-
-namespace EasilyNET.AutoDependencyInjection.Extensions;
-
-/// <summary>
-/// ApplicationInitialization扩展
-/// </summary>
-public static class ApplicationInitializationExtensions
-{
-    /// <summary>
-    /// 获取应用程序构建器
-    /// </summary>
-    /// <param name="applicationContext"></param>
-    /// <returns></returns>
-    // ReSharper disable once UnusedMethodReturnValue.Global
-    public static IApplicationBuilder GetApplicationBuilder(this ApplicationContext applicationContext) => applicationContext.ServiceProvider.GetRequiredService<IObjectAccessor<IApplicationBuilder>>().Value!;
-}
