+++ src/EasilyNET.AutoDependencyInjection/PropertyInjection/PropertyInjectionServiceProviderFactory.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
 
-namespace EasilyNET.PropertyInjection;
+namespace EasilyNET.AutoDependencyInjection.PropertyInjection;
 
 /// <summary>
 /// 属性注入服务提供者工厂
