+++ src/EasilyNET.MongoSerializer.AspNetCore/ServiceCollectionExtensions.cs
@@ -1,5 +1,4 @@
-﻿using Microsoft.Extensions.DependencyInjection;
-using MongoDB.Bson.Serialization;
+﻿using MongoDB.Bson.Serialization;
 using MongoDB.Bson.Serialization.Serializers;
 
 // ReSharper disable UnusedMethodReturnValue.Global
@@ -7,7 +6,7 @@ using MongoDB.Bson.Serialization.Serializers;
 // ReSharper disable UnusedMember.Global
 // ReSharper disable UnusedType.Global
 
-namespace EasilyNET.MongoSerializer.AspNetCore;
+namespace Microsoft.Extensions.DependencyInjection;
 
 /// <summary>
 /// 服务注册扩展类
