--- src/EasilyNET.PropertyInjection.AspNetCore/EasilyNET.PropertyInjection.AspNetCore.csproj
@@ -1,16 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-	<PropertyGroup>
-		<RepositoryUrl>https://github.com/EasilyNET/EasilyNET.PropertyInjection.AspNetCore</RepositoryUrl>
-		<Description>属性注入</Description>
-	</PropertyGroup>
-
-	<ItemGroup>
-		<FrameworkReference Include="Microsoft.AspNetCore.App" />
-	</ItemGroup>
-
-    <ItemGroup>
-      <ProjectReference Include="..\EasilyNET.PropertyInjection\EasilyNET.PropertyInjection.csproj" />
-    </ItemGroup>
-
-</Project>
