--- src/EasilyNET.PropertyInjection/EasilyNET.PropertyInjection.csproj
@@ -1,18 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-	<PropertyGroup>
-		<RepositoryUrl>https://github.com/EasilyNET/EasilyNET.PropertyInjection</RepositoryUrl>
-		<Description>属性注入</Description>
-	</PropertyGroup>
-
-	<ItemGroup>
-		<PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="8.0.0-preview.6.23329.7" />
-		<PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="8.0.0-preview.6.23329.7" />
-		<PackageReference Include="Microsoft.Extensions.Hosting.Abstractions" Version="8.0.0-preview.6.23329.7" />
-	</ItemGroup>
-
-	<ItemGroup>
-	  <ProjectReference Include="..\EasilyNET.Core\EasilyNET.Core.csproj" />
-	</ItemGroup>
-
-</Project>
