--- src/EasilyNET.PropertyInjection/README.md
@@ -1,1 +1,0 @@
-﻿#### EasilyNET.PropertyInjection
